Extract utility (date) function.
